requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.5.0
django-cloudinary-storage==0.3.0
django-import-export==4.3.7
django-unfold==0.51.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
urllib3==2.3.0
